Refactor TrusteeNotes to be an adapter for
Notes.tsx

Instead of inlining the configuratio for Notes.tsx in the TrusteeDetails

component we refactored the TrusteeNotes component to be an adapter

configuring Notes.tsx.

Jira ticket: CAMS-595
